requirements.txt
python-dotenv
streamlit
pandas
requests
requests
pandas
langchain_google_genai